pyproject.toml
[tool.poetry]
name = "correspodence-go"
version = "0.1.0"
description = "A command-line utility for playing Go"
authors = ["Leaf Foster Ascendant"]
readme = "README.md"
packages = [{include = "correspodence_go"}]

[tool.poetry.dependencies]
python = "^3.9"
pillow = "^10.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"

[tool.poetry.scripts]
corr-go = "correspodence_go.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
